Move numba functions to their own private file
Also move the try statement for importing choclo an numba to declutter
the `simulation.py` file. Fix some wording and typos in docstrings.
